requirements.txt
xgboost
scikit-learn
pandas
numpy
joblib
fastapi
uvicorn
anthropic
openai
python-dotenv
